add timeframe to state
