+++ src/modules/account/containers/account-overview-stats.ts
@@ -6,42 +6,9 @@ import * as constants from "modules/common-elements/constants";
 
 import AccountOverviewStats from "modules/account/components/account-overview-stats/account-overview-stats";
 
-const timeframeData = {
-	[constants.TIMEFRAMES.HOUR]: {
-		positions: 2,
-		marketsTraded: 5,
-		successfulDisputes: 9,
-		numberOfTrades: 50,
-		marketsCreated: 3,
-		redeemedPositions: 5,
-	},
-	[constants.TIMEFRAMES.DAY]: {
-		positions: 5,
-		marketsTraded: 10,
-		successfulDisputes: 19,
-		numberOfTrades: 67,
-		marketsCreated: 3,
-		redeemedPositions: 6,
-	},
-	[constants.TIMEFRAMES.WEEK]: {
-		positions: 24,
-		marketsTraded: 31,
-		successfulDisputes: 20,
-		numberOfTrades: 100,
-		marketsCreated: 9,
-		redeemedPositions: 10,
-	},
-	[constants.TIMEFRAMES.ALL]: {
-		positions: 100,
-		marketsTraded: 120,
-		successfulDisputes: 90,
-		numberOfTrades: 372,
-		marketsCreated: 39,
-		redeemedPositions: 51,
-	},
-};
-
-const mapStateToProps = state => ({});
+const mapStateToProps = state => ({
+	timeframeData: state.loginAccount.timeframeData
+});
 
 const mapDispatchToProps = dispatch => ({});
 
@@ -52,32 +19,32 @@ const mergeProps = (sP, dP, oP) => {
 	  	{
 	  		key: 0,
 	  		label: "Positions",
-	  		value: timeframeData[timeframe].positions,
+	  		value: sP.timeframeData[timeframe].positions,
 	  	},
 	  	{
 	  		key: 1,
 	  		label: "Number of Trades",
-	  		value: timeframeData[timeframe].marketsTraded,
+	  		value: sP.timeframeData[timeframe].marketsTraded,
 	  	},
 	  	{
 	  		key: 2,
 	  		label: "Markets Traded",
-	  		value: timeframeData[timeframe].successfulDisputes,
+	  		value: sP.timeframeData[timeframe].successfulDisputes,
 	  	},
 	  	{
 	  		key: 3,
 	  		label: "Markets Created",
-	  		value: timeframeData[timeframe].numberOfTrades,
+	  		value: sP.timeframeData[timeframe].numberOfTrades,
 	  	},
 	  	{
 	  		key: 4,
 	  		label: "Successful Disputes",
-	  		value: timeframeData[timeframe].marketsCreated,
+	  		value: sP.timeframeData[timeframe].marketsCreated,
 	  	},
 	  	{
 	  		key: 5,
 	  		label: "Redeemed Positions",
-	  		value: timeframeData[timeframe].redeemedPositions,
+	  		value: sP.timeframeData[timeframe].redeemedPositions,
 	  	}
 	];
