+++ src/modules/auth/actions/load-account-data.js
@@ -13,6 +13,7 @@ import { loadUserMarketTradingHistory } from "modules/markets/actions/market-tra
 import { loadDesignatedReporterMarkets } from "modules/reports/actions/load-designated-reporter-markets";
 import { loadDisputing } from "modules/reports/actions/load-disputing";
 import { loadGasPriceInfo } from "modules/app/actions/load-gas-price-info";
+import { updateTimeframeData } from "modules/account/actions/update-timeframe-data";
 
 const { ACCOUNT_TYPES } = augur.rpc.constants;
 
@@ -38,4 +39,5 @@ export const loadAccountData = (account, callback = logError) => dispatch => {
   dispatch(loadDesignatedReporterMarkets());
   dispatch(loadDisputing());
   dispatch(loadGasPriceInfo());
+  dispatch(updateTimeframeData());
 };
